FIX : 공고문 상세조회 카테고리 디자인 수정
