+++ src/pages/recruitDetails.jsx
@@ -1,15 +1,13 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate, useParams, useLocation } from 'react-router-dom';
 import backArrow from '../assets/images/backArrow.svg';
-import firstCategoryData from '../assets/categoryIndex/first_category.json';
-import secondCategoryData from '../assets/categoryIndex/second_category.json';
-import thirdCategoryData from '../assets/categoryIndex/third_category.json';
 import AlertModal from '../components/alertModal';
 import { postApplication } from '../api/application';
 import { UserStore } from '../store/userStore';
 import { closeRecruit, getRecruitDetail } from '../api/recruit';
 import SEO from '../components/seo';
 import { generateSeoContent } from '../utils/seo';
+import { getCategoryNames } from '../utils/categoryUtils.js';
 
 const parsePayment = (paymentString) => {
   if (!paymentString || typeof paymentString !== 'string') return null;
@@ -19,32 +17,6 @@ const parsePayment = (paymentString) => {
     num *= 10000;
   }
   return isNaN(num) ? null : num;
-};
-
-const getCategoryNames = (categoryDtoList) => {
-    if (!categoryDtoList || categoryDtoList.length === 0) {
-        return [];
-    }
-
-    return categoryDtoList.map(dto => {
-        const firstCatId = dto.firstCategory;
-        const secondCatId = dto.secondCategory;
-        const thirdCatId = dto.thirdCategory;
-
-        const firstName = firstCategoryData.first_category.find(
-            cat => cat.first_category_id === firstCatId
-        )?.name || '';
-
-        const secondName = secondCategoryData.second_category.find(
-            cat => cat.second_category_id === secondCatId
-        )?.name || '';
-
-        const thirdName = thirdCategoryData.third_category.find(
-            cat => cat.third_category_id === thirdCatId
-        )?.name || '';
-
-        return { first: firstName, second: secondName, third: thirdName };
-    });
 };
 
 export default function RecruitDetail() {
@@ -206,14 +178,6 @@ export default function RecruitDetail() {
     setIsApplySuccessModalOpen(true);
   };
 
-  // API에서 받은 상세 정보가 있으면 그것을 우선 사용, 없으면 기본 데이터 사용
-  // const displayData = recruitDetail || recruitData;
-  // const categoryList = recruitDetail?.categoryDtoList || recruitData?.categoryDtoList;
-  // const categoryNames = getCategoryNames(categoryList);
-
-  // const minPrice = recruitDetail ? parsePayment(recruitDetail.minPayment) : (displayData?.minPrice || null);
-  // const maxPrice = recruitDetail ? parsePayment(recruitDetail.maxPayment) : (displayData?.maxPrice || null);
-
   const displayData = recruitDetail;
   const categoryList = recruitDetail?.categoryDtoList || [];
   const categoryNames = getCategoryNames(categoryList);
@@ -308,28 +272,20 @@ export default function RecruitDetail() {
           </div>
           <h1 className="text-xl sm:text-3xl font-semibold">{displayData?.title}</h1>
           <div className="border-t border-gray-200 my-4 sm:my-6"></div>
-          
-          <div className="flex flex-col text-gray-600 mb-6 mt-2">
-            {categoryNames.map((category, index) => (
-              <div key={index} className="mb-1">
-                <span>{category.first}</span>
-                {category.second && (
-                  <>
-                <span className="mx-2">&gt;</span>
-                <span>{category.second}</span>
-                  </>
-                )}
-                {category.third && (
-                  <>
-                <span className="mx-2">&gt;</span>
-                <span className="font-medium text-black">{category.third}</span>
-                  </>
-                )}
-              </div>
-            ))}
-          </div>
           <div className="grid sm:grid-cols-2 gap-2 sm:gap-8 my-4 sm:my-6 text-sm sm:text-md">
             <div className="space-y-2 sm:space-y-4 ">
+              <div className="flex items-center flex-wrap mt-2">
+              <div className="flex flex-wrap gap-2">
+                {categoryNames.map((name, index) => (
+                  <span
+                    key={index}
+                    className="px-3 py-1 rounded-md bg-gray-100 text-gray-700 text-sm font-medium"
+                  >
+                    {name}
+                  </span>
+                ))}
+              </div>
+            </div>
               <div>
                 <span className="text-black mb-1">급여</span>
                 <span className="text-gray-500 mx-2">|</span>
